Refactor extract interface for UsernamePasswordExtractor
