+++ src/main/java/br/com/soeirosantos/poe/security/auth/login/extractor/AuthenticationExtractor.java
@@ -1,0 +1,10 @@
+package br.com.soeirosantos.poe.security.auth.login.extractor;
+
+import java.io.IOException;
+import javax.servlet.http.HttpServletRequest;
+import org.springframework.security.core.Authentication;
+
+public interface AuthenticationExtractor {
+
+    Authentication extract(HttpServletRequest request) throws IOException;
+}
